Fix ChannelRestControllerTest
